further refactoring and ui improvements
