+++ ui/package.json
@@ -8,6 +8,7 @@
     "bootstrap": "^5.2.0",
     "jwt-decode": "^3.1.2",
     "react": "^17.0.1",
+    "react-arborist": "^3.2.0",
     "react-bootstrap": "^2.5.0",
     "react-dom": "^17.0.1",
     "react-dropzone": "^11.3.2",
