+++ ui/public/index.html
@@ -4,11 +4,10 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
-    <meta name="description" content="RM FakeApi"
-    />
+    <meta name="description" content="RM FakeApi" />
     <title>rmfakecloud</title>
   </head>
-  <body>
+  <body style="background-color: #000">
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
   </body>
