+++ ui/src/pages/404.js
