+++ ui/src/pages/Connect/index.js
@@ -1,5 +1,5 @@
 import React, {useState, useLayoutEffect} from "react";
-import apiservice from "../services/api.service"
+import apiservice from "../../services/api.service"
 import Stack from 'react-bootstrap/Stack';
 import Button from 'react-bootstrap/Button';
 import { FaRepeat } from "react-icons/fa6";
