+++ ui/src/pages/Documents/Documents.css
@@ -1,0 +1,3 @@
+canvas {
+  margin: auto;
+}
