+++ ui/src/pages/Login/index.js
@@ -1,0 +1,66 @@
+import React, { useState } from "react";
+import { useAuthState } from "../../common/useAuthContext";
+import { loginUser } from "../../common/actions";
+import styles from "./Login.module.css";
+import { useHistory } from "react-router";
+import { Button, Form } from "react-bootstrap";
+
+const Login = () => {
+  let history = useHistory();
+  const [username, setUsername] = useState("");
+  const [password, setPassword] = useState("");
+
+  const { state, dispatch } = useAuthState(); //read the values of loading and errorMessage from context
+  const { errorMessage, loading } = state;
+
+  const handleLogin = async (e) => {
+    e.preventDefault();
+
+    let payload = { email: username, password };
+    try {
+      await loginUser(dispatch, payload);
+      history.push("/"); //TODO: usenavigate or return redirect
+    } catch (error) {
+      console.log(error);
+    }
+  };
+
+  return (
+    <div className={styles.container}>
+      <div style={{ width: 250 }}>
+        {errorMessage ? <p className={styles.error}>{errorMessage}</p> : null}
+
+        <Form>
+          <Form.Group className="mb-3">
+            <Form.Label htmlFor="username">Username</Form.Label>
+            <Form.Control
+              id="username"
+              value={username}
+              autoFocus
+              onChange={(e) => setUsername(e.target.value)}
+              disabled={loading}
+              placeholder="Username" />
+          </Form.Group>
+
+          <Form.Group className="mb-3">
+            <Form.Label htmlFor="password">Password</Form.Label>
+            <Form.Control
+              type="password"
+              id="password"
+              value={password}
+              onChange={(e) => setPassword(e.target.value)}
+              disabled={loading}
+              placeholder="Password" />
+          </Form.Group>
+
+          <Button type="submit" onClick={handleLogin} disabled={loading}>
+            Login
+          </Button>
+        </Form>
+
+      </div>
+    </div>
+  );
+};
+
+export default Login;
